labs/lab05-multi-agent/python/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
anthropic>=0.18.0
openai>=1.12.0
pydantic>=2.6.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests==2.31.0